Fix printing byte values in VK_EXT_host_image_copy test logs
Printing uint8_t values with iostream treats them as characters rather
than numbers, so they must be cast to a larger integer type first.

Affects:
dEQP-VK.image.host_image_copy.*

Components: Vulkan

+++ external/vulkancts/modules/vulkan/image/vktImageHostImageCopyTests.cpp
@@ -1090,8 +1090,8 @@ tcu::TestStatus HostImageCopyTestInstance::iterate(void)
                 if (data[i] != ((uint8_t *)resultBuffer.getDataPtr())[i])
                 {
                     log << tcu::TestLog::Message << "At byte " << i << " data from vkCopyImageToMemoryEXT() is "
-                        << data[i] << ", but data from vkCmdCopyImageToBuffer() (after padding) is "
-                        << ((uint8_t *)resultBuffer.getDataPtr())[i] << tcu::TestLog::EndMessage;
+                        << (int)data[i] << ", but data from vkCmdCopyImageToBuffer() (after padding) is "
+                        << (int)((uint8_t *)resultBuffer.getDataPtr())[i] << tcu::TestLog::EndMessage;
                     break;
                 }
             }
@@ -1614,8 +1614,8 @@ tcu::TestStatus PreinitializedTestInstance::iterate(void)
         {
             if (data[i] != ((uint8_t *)outputPtr)[i])
             {
-                log << tcu::TestLog::Message << "At byte " << i << " data from vkCopyImageToMemoryEXT() is " << data[i]
-                    << ", but data from vkCmdCopyImageToBuffer() is " << ((uint8_t *)outputPtr)[i]
+                log << tcu::TestLog::Message << "At byte " << i << " data from vkCopyImageToMemoryEXT() is "
+                    << (int)data[i] << ", but data from vkCmdCopyImageToBuffer() is " << (int)((uint8_t *)outputPtr)[i]
                     << tcu::TestLog::EndMessage;
                 break;
             }
